Better drag and drop
Drag and drop is now almost perfect from the node perspective. Unfortunately, the node only has access to the edges starting from itself, and not to the ones going towards it. That's why the second part of the edge positioning must be managed by a higher level object(and the node will probably have a reference to this one), probably the 'Graph' object.