Vector is now iterable
The vector class now implements iterable, which means it can be used in a foreach loop(and more). I did this because I hate the standard 'int i = 0' for loop; foreach is much more handy.
To do so I used a little 'hack'; Actually, the array inside the vector is converted into a list and the iterator is generated from that list. I did this because its much more handy that using array iterators themselves(and also much more handy than defining my own iterator).