Fixed initial node flickering
Fixed a bug where the node would flicker a bit at first when dragged. Appartently it was due to the distance text, which was stored in the node's VBox.
The distance text has been removed and the distance management will be delegated to another class when implementing the BFM algorithm.